+++ tests/testthat/test-rdb.R
@@ -31,37 +31,37 @@ test_that("Nothing provided", {
 })
 
 test_that("api_link length is greater than one or equal to zero", {
-  expect_error(rdb(url = c("url1", "url2")))
-  expect_error(rdb(url = character()))
+  expect_error(rdb(api_link = c("url1", "url2")))
+  expect_error(rdb(api_link = character()))
 })
 
 test_that("api_link is not a character string", {
-  expect_error(rdb(url = 0))
+  expect_error(rdb(api_link = 0))
 })
 
 test_that("use_readLines is NULL", {
-  expect_error(rdb(url = "url", use_readLines = NULL))
+  expect_error(rdb(api_link = "url", use_readLines = NULL))
 })
 
 test_that("use_readLines length is greater than one or equal to zero", {
-  expect_error(rdb(url = "url", use_readLines = c(TRUE, FALSE)))
-  expect_error(rdb(url = "url", use_readLines = logical()))
+  expect_error(rdb(api_link = "url", use_readLines = c(TRUE, FALSE)))
+  expect_error(rdb(api_link = "url", use_readLines = logical()))
 })
 
 test_that("use_readLines is not a logical", {
-  expect_error(rdb(url = "url", use_readLines = "TRUE"))
+  expect_error(rdb(api_link = "url", use_readLines = "TRUE"))
 })
 
 test_that("curl_config is not a curl_handle object or a named list", {
-  expect_error(rdb(url = "url", curl_config = TRUE))
-  expect_error(rdb(url = "url", curl_config = list(1)))
+  expect_error(rdb(api_link = "url", curl_config = TRUE))
+  expect_error(rdb(api_link = "url", curl_config = list(1)))
 })
 
 test_that("filters is not a valid list", {
-  expect_error(rdb(url = "url", filters = TRUE))
-  expect_error(rdb(url = "url", filters = "filters"))
-  expect_error(rdb(url = "url", filters = list()))
-  expect_error(rdb(url = "url", filters = list(code = "interpolate", parameter = NULL)))
-  expect_error(rdb(url = "url", filters = list(code = "interpolate", parameters = list())))
-  expect_error(rdb(url = "url", filters = list(list(code = "interpolate", parameters = list(frequency = "quarterly", method = "spline")), list(code = "aggregate", parameters = list()))))
+  expect_error(rdb(api_link = "url", filters = TRUE))
+  expect_error(rdb(api_link = "url", filters = "filters"))
+  expect_error(rdb(api_link = "url", filters = list()))
+  expect_error(rdb(api_link = "url", filters = list(code = "interpolate", parameter = NULL)))
+  expect_error(rdb(api_link = "url", filters = list(code = "interpolate", parameters = list())))
+  expect_error(rdb(api_link = "url", filters = list(list(code = "interpolate", parameters = list(frequency = "quarterly", method = "spline")), list(code = "aggregate", parameters = list()))))
 })